spike-37: address some todo comments
