+++ nest-app/src/validators/validators.controller.ts
@@ -1,11 +1,10 @@
 import { Controller, Get, Query } from "@nestjs/common";
 import { Transform } from "class-transformer";
 import { IsIn, IsNumber, IsOptional, IsString } from "class-validator";
-import { ValidatorProcessed } from "src/types/api";
+import { Sort, ValidatorProcessed } from "src/types/api";
 
-import { Sort, ValidatorsService } from "./validators.service";
+import { ValidatorsService } from "./validators.service";
 
-// TODO: maybe this should be used for all pagination requests?
 export class ValidatorsQueryDtp {
   @Transform(({ value }) => parseInt(value, 10))
   @IsNumber()
